added walletCard to WalletTracker and few bugfixes

+++ src/views/user/WalletTracker.vue
@@ -43,47 +43,11 @@
                     <wallet-chart :wallets='wallets'></wallet-chart>
                 </v-flex>
             </div>
-            <app-card
-                    colClasses="xl12"
-                    customClasses="transparent no-box-shadow"
-                    :fullScreen="true"
-                    :fullBlock="true"
-                    :footer="true"
-                    v-if="wallets.length>0"
-
-                >
-                <div class='assets-wrapper'>
-                    <app-section-loader :status="loader"></app-section-loader>
-                    <v-card v-for="(wallet, index) in wallets" v-bind:key="index" class='assets d-md-full'>
-                        <div class="assets-item">
-                            <div class="assets-preview">
-                                <div class="assets-preview__text">{{wallet.preview}}</div>
-                            </div>
-                            <div class="assets-label">{{wallet.label}}
-
-                            <v-btn
-                            color="red"
-                            dark
-                            small
-                            absolute
-                            top
-                            right
-                            fab
-                            @click='removeWallet(wallet)'
-                            >
-                                <v-icon>delete_outline</v-icon>
-                            </v-btn>
-                            </div>
-                            <div class="assets-balance">
-                                <div class="assets-balance__wallet">{{wallet.address}}</div>
-                                <div class="assets-balance__label">{{$t('message.balance')}}</div>
-                                <div class="assets-balance__item">{{wallet.balance}} NKN</div>
-                                <div class="assets-balance__item">≈${{wallet.balanceUsd}}</div>
-                            </div>
-                        </div>
-                    </v-card>
-                </div>
-            </app-card>
+            <v-layout row wrap v-if="wallets.length>0">
+                <v-flex xs4 v-for="(wallet, index) in wallets" v-bind:key="index">
+                    <wallet-card :address="wallet.address" :deleteCallback="removeWallet"></wallet-card>
+                </v-flex>
+            </v-layout>
 
      </v-container>
 </div>
@@ -92,12 +56,15 @@
 import axios from "axios";
 import MarketStats from "Components/Widgets/MarketStats";
 import WalletChart from "Components/Widgets/WalletChart";
+import WalletCard from "Components/WalletCard/WalletCard";
+
 
 export default {
     components: {
+        WalletCard,
         MarketStats,
         WalletChart
-  },
+    },
     data() {
         return {
             interval: null,
@@ -141,14 +108,22 @@ export default {
            self.isMultiError = error.response.data.msg
         })
     },
-    removeWallet(wallet) {
+    removeWallet(walletAddr) {
+        let wallet = null;
+        this.wallets.forEach(function(singleWallet) {
+             if (singleWallet.address == walletAddr){
+                 wallet = singleWallet;
+             }
+        });
         const self = this;
-        let id = wallet.id
-        axios.delete('walletAddresses/'+id, {
-                })
-        .then((response) => {
-            this.getWalletsBalance()
-        })
+        if(wallet){
+            let id = wallet.id
+            axios.delete('walletAddresses/'+id, {
+                    })
+            .then((response) => {
+                this.getWalletsBalance()
+            })
+        }
     },
     getWalletsBalance(){
         const self = this;
